# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SI GAPLEK - Sistem Gudang dan Pengelolaan Logistik Kantor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opname:
    get:
      operationId: listOpname
      tags: [opname]
      summary: List stock opname sessions
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opname"
    post:
      operationId: createOpname
      tags: [opname]
      summary: Create stock opname session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpnameBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opname"
  /opname/{id}:
    get:
      operationId: getOpname
      tags: [opname]
      summary: Get opname session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opname
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpnameDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opname/{id}/finalize:
    post:
      operationId: finalizeOpname
      tags: [opname]
      summary: Finalize opname session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opname"
  /reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      summary: Stock report
      parameters:
        - name: warehouseId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Stock report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockReportItem"
  /reports/transactions:
    get:
      operationId: getTransactionReport
      tags: [reports]
      summary: Transaction report
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Transaction report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionReportItem"
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit-logs]
      summary: List audit logs
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Opname:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        warehouseId:
          type: ["integer", "null"]
        warehouseName:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - referenceNo
        - status
        - startDate
        - createdAt
    OpnameDetail:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        warehouseId:
          type: ["integer", "null"]
        warehouseName:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OpnameItem"
      required:
        - id
        - referenceNo
        - status
        - startDate
        - createdAt
        - items
    OpnameItem:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        systemQty:
          type: integer
        actualQty:
          type: ["integer", "null"]
        difference:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - itemId
        - itemCode
        - itemName
        - systemQty
    CreateOpnameBody:
      type: object
      properties:
        warehouseId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        startDate:
          type: string
        itemIds:
          type: array
          items:
            type: integer
      required:
        - startDate
        - itemIds
    StockReportItem:
      type: object
      properties:
        itemId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        categoryName:
          type: ["string", "null"]
        unitName:
          type: ["string", "null"]
        currentStock:
          type: integer
        minimumStock:
          type: integer
        unitPrice:
          type: number
        totalValue:
          type: number
        status:
          type: string
      required:
        - itemId
        - itemCode
        - itemName
        - currentStock
        - minimumStock
        - unitPrice
        - totalValue
        - status
    TransactionReportItem:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        type:
          type: string
        status:
          type: string
        totalItems:
          type: integer
        transactionDate:
          type: string
        createdByName:
          type: ["string", "null"]
      required:
        - id
        - referenceNo
        - type
        - status
        - totalItems
        - transactionDate
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        action:
          type: string
        description:
          type: string
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - entityType
        - action
        - description
        - createdAt
...